Properly register entity transform event depending on version
